REF: move interpolate validation to core.missing (#53580)
* REF: move interpolate validation to core.missing

* lint fixup
